requirements.txt
requests
django
python-dotenv
djangorestframework
psycopg2-binary
